backend/requirements.txt
flask
pymongo
flask-cors
python-dotenv
flask-jwt-extended
bcrypt
numpy
opencv-python
PyJWT
werkzeug
dlib
face_recognition
imutils
Pillow
